% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_HICS.R
\name{calculate_HICS}
\alias{calculate_HICS}
\title{Calculate highest intensity charge state}
\usage{
calculate_HICS(sequence, height = 4)
}
\arguments{
\item{sequence}{Sequence of proteoform to find HICS for. Can only include the 20
canonical amino acids.}

\item{height}{Height to use for cutting tree generated by hierarchical clustering.}
}
\value{
A numeric vector of the theoretical HICS for the provided sequence.
}
\description{
Calculates the theoretical highest intensity charge state based on a method
inspired by Douglass and Venter (dx.doi.org/10.1021/ac401245r).
}
\examples{
calculate_HICS("AAAAKAAAA", height = 4)
}
