% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot}
\alias{cplot}
\title{Scatter plot with colour as a third dimension}
\usage{
cplot(x, y, z, cols, legend.title = "legend.title", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Scatter plot with colour as a third dimension
}
