% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplotShort.R
\name{cplotShort}
\alias{cplotShort}
\title{Scatter plot with colour as a third dimension (no legend, for small plots)}
\usage{
cplotShort(x, y, z, cols, zr, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Scatter plot with colour as a third dimension (no legend, for small plots)
}
