% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpoints.R
\name{cpoints}
\alias{cpoints}
\title{Scatter points with colour as a third dimension}
\usage{
cpoints(x, y, z, cols, legend.title, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Scatter points with colour as a third dimension
}
