% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpointsTruth.R
\name{cpointsTruth}
\alias{cpointsTruth}
\title{Helper function for diagonal panel in pairs()
plot points in colour palette, and the last point in a different colour}
\usage{
cpointsTruth(x, y, z, cols, truecol = "green", legend.title, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Helper function for diagonal panel in pairs()
plot points in colour palette, and the last point in a different colour
}
