% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfunc.R
\name{dfunc}
\alias{dfunc}
\title{Add 2d kernel density estimates to pairs() plot.}
\usage{
dfunc(x, y, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add 2d kernel density estimates to pairs() plot.
}
