% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfunc_up.R
\name{dfunc_up}
\alias{dfunc_up}
\title{Add density plot(with specified colours) to a pairs plot}
\usage{
dfunc_up(x, y, dfunc.col = grey.colors(10), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add density plot(with specified colours) to a pairs plot
}
