% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfunc_up_truth.R
\name{dfunc_up_truth}
\alias{dfunc_up_truth}
\title{plotting 2d kernel density estimates in pairs() plot,
adding a data point overlay in the last row of the input matrix.}
\usage{
dfunc_up_truth(
  x,
  y,
  dfunc.col = grey.colours(10),
  col = "black",
  bg = "red",
  cex = 1.5,
  pch = 21,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plotting 2d kernel density estimates in pairs() plot,
adding a data point overlay in the last row of the input matrix.
}
