% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_intersection_degree_plot.R
\name{make_intersection_degree_plot}
\alias{make_intersection_degree_plot}
\title{Make an intersection degree plot from a set of unique identifiers}
\usage{
make_intersection_degree_plot(
  IDlist,
  Yrange = c(0, 100),
  plotType = "Protein",
  fillColor = "#4C4184",
  fontFamily = "sans"
)
}
\arguments{
\item{IDlist}{A list of vectors. Each list item should be a vector of unique
identifiers corresponding to protein or proteoform IDs, e.g. UniProt accession
numbers or proteoform record numbers from the Consortium for Top-Down Proteomics
proteoform atlas.}

\item{Yrange}{Percentage range to use for the Y-axis of the plot. Defaults to `c(0,100)`.}

\item{plotType}{Type of plot to create, typically "Protein" or "Proteoform".This only
affects the Y-axis title and can be any string.}

\item{fillColor}{Fill color to use for bars on plot. Should be specified as a Hex code.
Defaults to MagLab violet (#4C4184).}

\item{fontFamily}{Font family to use for plot. Defaults to "sans".}
}
\value{
An intersection degree plot (a ggplot object).
}
\description{

}
\examples{

IDlist <-
  list(
       "Frac. 1" = c("A", "B", "C", "D"),
       "Frac. 2" = c("A", "B", "E"),
       "Frac. 3" = c("A", "E", "F")
      )

make_intersection_degree_plot(IDlist)


}
