% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSet_maker.R
\name{make_UpSet_plot}
\alias{make_UpSet_plot}
\title{Make an UpSet plot from a list of unique identifiers}
\usage{
make_UpSet_plot(
  UpSetlist,
  plotType = "Protein",
  savePDF = FALSE,
  outputDir = getwd()
)
}
\arguments{
\item{UpSetlist}{A list of lists of identifiers properly formatted for use by the
UpSetR::upset() function.}

\item{plotType}{Type of UpSet plot to make. This only affects the axis
titles and filename. Typical values are "Protein" or "Proteoform". Defaults to "Protein".}

\item{savePDF}{Boolean value, controls whether to save PDF output to outputDir. Defaults to FALSE.}

\item{outputDir}{Directory to save PDF output. Defaults to R working directory.}
}
\value{
An UpSet plot created by UpSetR::upset(). If savePDF is TRUE, a PDF is saved
to the output directory.
}
\description{
make_UpSet_plot() creates an UpSet plot from a list of unique identifiers such as
UniProt accession numbers or proteoform record numbers.
}
\examples{
make_UpSet_plot(
   list(
      list(
         "Fraction1" = c("A", "B", "C", "D", "E"),
         "Fraction2" = c("C", "D", "E", "F", "G")
      )
   )
)

}
