% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_hit_report.R
\name{modify_hit_report}
\alias{modify_hit_report}
\title{modify_hit_report}
\usage{
modify_hit_report(
  hitReport,
  isoformReport = NULL,
  proteoformReport = NULL,
  IDtype = "protein"
)
}
\arguments{
\item{hitReport}{A data frame containing a hit report produced by TDViewer.}

\item{isoformReport}{A data frame containing an isoform report produced by TDViewer.}

\item{proteoformReport}{A data frame containing a proteoform report produced by TDViewer.}

\item{IDtype}{Type of ID in the hit report to filter by, either "protein" or "proteoform".}
}
\value{

}
\description{
modify_hit_report
}
