% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaatLinePlot.R
\name{oaatLinePlot}
\alias{oaatLinePlot}
\title{Line plot for one-at-a-time sensitivity analysis}
\usage{
oaatLinePlot(
  X_oaat,
  y_oaat_mean,
  y_oaat_sd,
  n_oaat,
  nr,
  nc,
  linecol = "black",
  shadecol = "grey",
  ...
)
}
\arguments{
\item{X_oaat}{Matrix of one-at-time input design generated by oaat_design in package emtools}

\item{y_oaat_mean}{vector of one-at-a-time output}

\item{y_oaat_sd}{standard deviation vector of one-at-a-time output.}

\item{n_oaat}{Number of samples for each input}

\item{nr}{Number of rows in the plot}

\item{nc}{Number of colums in the plot}

\item{linecol}{line colour}

\item{shadecol}{uncertainty shade colour}

\item{...}{Further plotting parameters (input to plot)}
}
\value{
A plot of the impact of turning each parameter from low to high in isolation.
}
\description{
Line plot for one-at-a-time sensitivity analysis
}
