% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaatLinePlotMulti.R
\name{oaatLinePlotMulti}
\alias{oaatLinePlotMulti}
\title{Title}
\usage{
oaatLinePlotMulti(X_oaat, Y_oaat, n_oaat, nr, nc, ...)
}
\arguments{
\item{X_oaat}{Matrix of one-at-time input design generated by oaat_design in package emtools}

\item{Y_oaat}{Normalised matrix of one-at-a-time output, with variables in the columns and rows matching X_oaat}

\item{n_oaat}{Number of samples for each input}

\item{nr}{Number of rows in the plot}

\item{nc}{Number of colums in the plot}

\item{...}{}
}
\value{
A plot that overlays multiple one-at-a-time sensitivity estimates. Shading is not included for clarity.
}
\description{
Title
}
