% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_aug.R
\name{pairs_aug}
\alias{pairs_aug}
\title{Augment pairs plots with different colours}
\usage{
pairs_aug(..., col = c("grey", "red"), pch = c(21, 19))
}
\arguments{
\item{pch}{}
}
\value{

}
\description{
Augment pairs plots with different colours
}
