% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_HICS_hclust.R
\name{predict_HICS_hclust}
\alias{predict_HICS_hclust}
\title{Predict the highest intensity charge state using hierarchical clustering}
\usage{
predict_HICS_hclust(sequence, height = 3)
}
\arguments{
\item{sequence}{Vector of sequences of proteoforms to find HICS for. Can only include the 20
canonical amino acids.}

\item{height}{Height to use for cutting tree generated by hierarchical clustering. Effectively, max distance between amino acids considered to be in a single charge "group".}
}
\value{
A numeric vector of the theoretical HICS for the provided sequence.
}
\description{
Calculates the theoretical highest intensity charge state based on a method
adapted from Douglass and Venter (dx.doi.org/10.1021/ac401245r).
}
\examples{
predict_HICS_hclust("AAAAKAAAA", height = 3)
}
