% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSet_maker.R
\name{UpSet_maker}
\alias{UpSet_maker}
\title{Make an UpSet plot from a list of unique identifiers}
\usage{
UpSet_maker(
  UpSetlist,
  UpSetType = "protein",
  savePDF = FALSE,
  outputDir = getwd()
)
}
\arguments{
\item{UpSetlist}{A list of lists of identifiers properly formatted for use by the
UpSetR::upset() function. This can be provided by dissect_TDsummary().}

\item{UpSetType}{Type of UpSet plot to make. This only affects the axis
titles. Possible values are "protein" or "proteoform".}
}
\value{
An UpSet plot created by UpSetR::upset(). If savePDF is TRUE, a PDF is saved
to the output directory.
}
\description{
UpSet_maker() creates an UpSet plot from a list of unique identifiers such as
UniProt accession numbers or proteoform record numbers.
}
\examples{
UpSet_maker(
   list(
      list(
         "Fraction1" = c("A", "B", "C", "D", "E"),
         "Fraction2" = c("C", "D", "E", "F", "G")
      )
   )
)

}
