% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_iron.R
\name{waffle_iron}
\alias{waffle_iron}
\title{Make a waffle plot}
\usage{
waffle_iron(
  df,
  shortName = NULL,
  savePDF = FALSE,
  outputDir = getwd(),
  shortname_colname = "Short name",
  fraction_colname = "fraction",
  loc_colnames = c("Cytosolic Proteins", "Membrane Proteins", "Periplasmic Proteins",
    "NOTA Proteins"),
  waffleType = "Protein"
)
}
\arguments{
\item{df}{A data frame containing columns for shortname, fraction, and localization counts columns.}

\item{shortName}{Shortname to use for making waffle plot.}

\item{savePDF}{Boolean value, controls whether to save PDF output to outputDir. Defaults to FALSE.}

\item{outputDir}{Directory to save PDF output. Defaults to R working directory.}

\item{shortname_colname}{Name of data frame column containing shortnames. Defaults to 'Short name'.}

\item{fraction_colname}{Name of data frame column containing fractions. Defaults to 'fraction'.}

\item{loc_colnames}{Vector of names of data frame columns containing counts of
cytosolic, membrane, periplasmic, and none-of-the-above/unannotated proteins. Defaults to
c("Cytosolic Proteins", "Membrane Proteins", "Periplasmic Proteins", "NOTA Proteins").}

\item{waffleType}{Type of Waffle plot to make. This only affects the axis
titles and filename. Typical values are "Protein" or "Proteoform". Defaults to "Protein".}
}
\value{

}
\description{
Make a waffle plot from a data frame containing columns for short name,
fraction, and subcellular localizations.
}
