% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_intersect_plot.R
\name{make_intersect_plot}
\alias{make_intersect_plot}
\title{Make an intersect plot for a set of protein/proteoform IDs}
\usage{
make_intersect_plot(
  IDlist,
  shortNames,
  mediumUsed,
  fracMethod,
  Yrange = c(0, 100),
  plotType = "Protein"
)
}
\arguments{
\item{plotType}{}
}
\value{

}
\description{
Make an intersect plot for a set of protein/proteoform IDs
}
