% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_iron.R
\name{waffle_iron}
\alias{waffle_iron}
\title{Make a waffle plot}
\usage{
waffle_iron(
  df,
  savePDF = FALSE,
  outputDir = getwd(),
  outputPrefix = format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S"),
  fraction_colname = "fraction",
  waffleType = "Protein"
)
}
\arguments{
\item{df}{A data frame containing columns for fraction and subcellular localization counts.}

\item{savePDF}{Boolean value, controls whether to save PDF output to outputDir. Defaults to FALSE.}

\item{outputDir}{Directory to save PDF output. Defaults to R working directory.}

\item{outputPrefix}{String to use to prepend heatmap output filenames.
Defaults to date and time.}

\item{fraction_colname}{Name of data frame column containing fractions. Defaults to 'fraction'.}

\item{waffleType}{Type of Waffle plot to make. This only affects the axis
titles and filename. Typical values are "Protein" or "Proteoform". Defaults to "Protein".}
}
\value{
A waffle plot (ggplot object)
}
\description{
Make a waffle plot from a data frame containing columns for
fraction and subcellular localization.
}
\examples{

df <-
   data.frame(
      "fraction" =
         c(1,2,3,4,5,6),
      "Cytosol" =
         c(20,25,30,30,20,10),
      "Membrane" =
         c(15,10,20,25,10,5)
   )

   waffle_iron(
      df
   )
}
