% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_iron.R
\name{waffle_iron}
\alias{waffle_iron}
\title{Make a waffle plot}
\usage{
waffle_iron(
  df,
  shortName = NULL,
  savePDF = TRUE,
  outputDir = getwd(),
  shortname_colname = "Short name",
  fraction_colname = "fraction",
  loc_colnames = c("Cytosolic Proteins", "Membrane Proteins", "Periplasmic Proteins",
    "NOTA Proteins"),
  waffleType = "Protein"
)
}
\arguments{
\item{waffleType}{}
}
\value{

}
\description{
Make a waffle plot from a data frame containing columns for short name,
fraction, and subcellular localizations.
}
